Dark Line now with 2D Pointer, but DarkLine is not stored in 2D Pointer

+++ 3_Programme/OpenCV_MSVC/darkline_pointer.h
@@ -6,16 +6,16 @@
 #include <mutex>
 #include <thread>
 
-// Structure to hold image data with double pointer
+// 图像数据结构，使用 double 类型的 2D pointer
 struct ImageData {
-    uint16_t** data;  // 2D double pointer for image data
+    double** data;    // 2D pointer for image data
     int rows;         // Number of rows in the image
     int cols;         // Number of columns in the image
 
     // Constructor
     ImageData() : data(nullptr), rows(0), cols(0) {}
 
-    // Destructor to clean up memory
+    // Destructor
     ~ImageData() {
         if (data) {
             for (int i = 0; i < rows; i++) {
@@ -28,6 +28,7 @@ struct ImageData {
 
 class DarkLinePointerProcessor {
 public:
+    // 暗线结构体
     struct DarkLine {
         int x;           // x coordinate of the dark line
         int y;           // y coordinate for horizontal lines
@@ -41,22 +42,18 @@ public:
     };
 
     enum class RemovalMethod {
-        NEIGHBOR_VALUES,  // Using neighboring values to fill
+        NEIGHBOR_VALUES,  // Using neighbor values to fill
         DIRECT_STITCH    // Direct stitching
     };
 
-    // Main functionality modified to use ImageData
+    // Main functionality
     static std::vector<DarkLine> detectDarkLines(const ImageData& image);
     static void removeDarkLines(ImageData& image, const std::vector<DarkLine>& lines);
 
-    static void removeAllDarkLines(ImageData& image,
-                                   std::vector<DarkLine>& detectedLines);
-
-    static void removeInObjectDarkLines(ImageData& image,
-                                        std::vector<DarkLine>& detectedLines);
-
-    static void removeIsolatedDarkLines(ImageData& image,
-                                        std::vector<DarkLine>& detectedLines);
+    // Specialized removal functions
+    static void removeAllDarkLines(ImageData& image, std::vector<DarkLine>& detectedLines);
+    static void removeInObjectDarkLines(ImageData& image, std::vector<DarkLine>& detectedLines);
+    static void removeIsolatedDarkLines(ImageData& image, std::vector<DarkLine>& detectedLines);
 
     static void removeDarkLinesSelective(
         ImageData& image,
@@ -65,17 +62,19 @@ public:
         bool removeIsolated,
         RemovalMethod method = RemovalMethod::NEIGHBOR_VALUES);
 
-    static uint16_t findReplacementValue(
+    // Helper functions for pixel value replacement
+    static double findReplacementValue(
         const ImageData& image,
         int x, int y,
         bool isVertical,
         int lineWidth);
 
-    static uint16_t findStitchValue(
+    static double findStitchValue(
         const ImageData& image,
         const DarkLine& line,
         int x, int y);
 
+    // Sequential line removal
     static void removeDarkLinesSequential(
         ImageData& image,
         const std::vector<DarkLine>& lines,
@@ -85,14 +84,15 @@ public:
 
 private:
     // Constants for detection and removal
-    static constexpr uint16_t BLACK_THRESHOLD = 1000;
-    static constexpr float NOISE_TOLERANCE = 0.1f;
+    static constexpr double BLACK_THRESHOLD = 1000.0;
+    static constexpr double NOISE_TOLERANCE = 0.1;
     static constexpr int MIN_LINE_WIDTH = 1;
     static constexpr int VERTICAL_CHECK_RANGE = 3;
     static constexpr int HORIZONTAL_CHECK_RANGE = 2;
-    static constexpr uint16_t WHITE_THRESHOLD = 55000;
-    static constexpr uint16_t MIN_BRIGHTNESS = 1000;
+    static constexpr double WHITE_THRESHOLD = 55000.0;
+    static constexpr double MIN_BRIGHTNESS = 1000.0;
 
+    // Helper functions
     static int calculateSearchRadius(int lineWidth);
     static bool isInObject(const ImageData& image, int pos, int lineWidth, bool isVertical, int threadId);
 };
